config-service/pyproject.toml
[project]
name = "config-service"
version = "0.1.0"
description = "Centralized configuration management REST API"
requires-python = ">=3.13"

dependencies = [
    "fastapi==0.116.1",
    "uvicorn==0.34.0",
    "psycopg2-binary==2.9.10",
    "python-ulid==3.0.0",
    "pydantic-settings==2.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.4.1",
    "httpx==0.28.1",
    "pytest-asyncio==0.25.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["app"]
python_files = ["*_test.py"]
python_classes = []
python_functions = ["test_*"]
